Implement some services for user

+++ src/main/java/vn/jason/auth/user/UserRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import java.util.Set;
 
 public interface UserRepository {
-    Set<User> getAll(int page, int recordsPerPage);
+    Set<User> getAll(long page, long recordsPerPage);
     Optional<User> get(long id);
     Optional<User> add(User user);
 }
